+++ src/Exception/Exception.php
@@ -2,6 +2,6 @@
 
 namespace TomPHP\ConfigServiceProvider\Exception;
 
-class RuntimeException extends \RuntimeException
+interface Exception
 {
 }