+++ src/Exception/InvalidConfigException.php
@@ -2,6 +2,6 @@
 
 namespace TomPHP\ConfigServiceProvider\Exception;
 
-class InvalidConfigException extends RuntimeException
+class InvalidConfigException extends \RuntimeException implements Exception
 {
 }