+++ tests/Exception/EntryDoesNotExistExceptionTest.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace tests\TomPHP\ConfigServiceProvider\Exception;
+
+use PHPUnit_Framework_TestCase;
+use TomPHP\ConfigServiceProvider\Exception\EntryDoesNotExistException;
+
+final class EntryDoesNotExistExceptionTest extends PHPUnit_Framework_TestCase
+{
+    public function testItImplementsTheBaseExceptionType()
+    {
+        $this->assertInstanceOf(
+            'TomPHP\ConfigServiceProvider\Exception\Exception',
+            new EntryDoesNotExistException()
+        );
+    }
+
+    public function testItIsADomainException()
+    {
+        $this->assertInstanceOf(
+            'TomPHP\ConfigServiceProvider\Exception\Exception',
+            new EntryDoesNotExistException()
+        );
+    }
+}