+++ tests/Exception/UnknownFileTypeExceptionTest.php
@@ -7,10 +7,10 @@ use TomPHP\ConfigServiceProvider\Exception\UnknownFileTypeException;
 
 final class UnknownFileTypeExceptionTest extends PHPUnit_Framework_TestCase
 {
-    public function testItIsAnInstanceOfThePackagesRuntimeException()
+    public function testItImplementsTheBaseExceptionType()
     {
         $this->assertInstanceOf(
-            'TomPHP\ConfigServiceProvider\Exception\RuntimeException',
+            'TomPHP\ConfigServiceProvider\Exception\Exception',
             new UnknownFileTypeException()
         );
     }